Update upload/catalog/controller/payment/authorizenet_aim.php
Added extra fields for shipping information, as Authorize.Net was rejecting the Order Confirmation with error "Ship To First name is required". Couple of clients also reported this issue. Don't know if it is fixed in any other way already. But I have reporoduced this issue on latest version and error coming up. So I am proposing this change and its working fine now.