fix for:
When turning on SEO Urls on a clean install of 1.5.5.1 I'm getting the following error when trying to go to the homepage.

Fatal error: Call to a member function query() on a non-object in *My Website*/public_html/catalog/controller/common/seo_url.php on line 74

Sorry if this is not a bug, but after looking around I could not find this mentioned anywhere else.

If it helps, I'm running PHP Version 5.3.19 with safe mode off.