Fix label issue (incorrect attribute)
Add order status defaults instead of all defaulting to voided. User will need to change if custom order status are used but will save time for those with defaults.

Changed the braintree setup url on opencart to final config. It is configured for staging setup but will be swapped to live when module is launched.