Added checks around each error part that could be passed from PayPal if it is post curl request. If no error is found then a generic error is displayed to the user to advise contacting the store owner. Without any php notice being thrown.
The system was made originally with an EXTREMELY accurate log collection around almost every PayPal call and action. Simple solution is to check the logs - enable them in the PayPal module then look at them AFTER you can replicate the problem.

Blindly creating a bug report over a PHP notice is not helpful, logs are helpful so please next time provide the actual error that PayPal says which will be found in your log. Worst case scenario is that PayPal sandbox is having problems - I have created many issues before only to find out it was a 5 minute window causing problems on the sandbox and even live checkout systems too.

https://developer.paypal.com/support/ - They can advise what error message PayPal gave and why.

As Express checkout is now ageing there is also the chance of legacy code being phased out - we usually get informed prior to it happening but if that is reported by PayPal tech support please create another issue for it.

Close #6221