Payflow Pro gateway is now legacy but docs are still maintained at PayPal. Module was built years ago.
As per issue #5875 - start date was being passed yet it is not even part of the PayPal NVPs so is most likely the cause of a constant fail transaction. Unable to obtain test info for this system as it is older but these updates can also port backwards if needed (depending how far back the mistake goes and who did it).

The pp payflow iframe controller also contained multiple incorrect config references (referencing the pp pro iframe module) - so again this would not have worked as expected either.

Both modules should now be resolved with unable to test without credentials. Please re-open with log info if continues. PayPal technical support will also be able to offer advice as to the cause of transactions not going through - as it could also be related to the account setup/config.

Close #5875