Resolved the issue with PayPal JS script being loaded improperly at the ajax stage of the checkout causing button rendering to fail.
Added OC Event to add the PayPal JS script to load with the default checkout/checkout page to ensure that the JS file is preloaded and available for event call if PayPal is selected as the payment option.

Added check in the settings to ensure that the event is installed if OC was upgraded and PP Express was already enabled.